Custom role permission added with seeder and test case

+++ src/__test__/setup.js
@@ -1,5 +1,6 @@
 const { MongoMemoryServer } = require('mongodb-memory-server');
 const mongoose = require('mongoose');
+const { seed } = require('../seeder/databaseSeeder');
 
 let mongod;
 beforeAll(async () => {
@@ -7,12 +8,18 @@ beforeAll(async () => {
   const mongoUri = await mongod.getUri();
 
   await mongoose.connect(mongoUri);
+
+  // Seed the database
+  await seed();
 });
 
 beforeEach(async () => {
   const collections = await mongoose.connection.db.collections();
   for (let collection of collections) {
-    await collection.deleteMany({});
+    // Clear the database except for the module, roles and permissions collection
+    if (!['modules', 'roles', 'permissions'].includes(collection.collectionName)) {
+      await collection.deleteMany({});
+    }
   }
 });
